my_blog/requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-appconf==1.0.3
django-ckeditor==5.8.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-taggit==1.2.0
idna==2.8
jsonfield==2.0.2
Markdown==3.1.1
oauthlib==3.1.0
pilkit==2.0
Pillow==7.0.0
pycparser==2.19
Pygments==2.5.2
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
